bot/requirements.txt
python-telegram-bot[job-queue]==20.7
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
psutil==5.9.6
